[Refactor] OCR 로직 분리 (#66)
