+++ BookKitty/BookKitty/BookMatchKit/Sources/BookMatchKit/Protocols/TextExtractable.swift
@@ -1,0 +1,6 @@
+import RxSwift
+import UIKit
+
+public protocol TextExtractable {
+    func extractText(from image: UIImage) -> Single<[String]>
+}
